Move broadcasting related functions to broadcast.rs
